1.8.0版本跟踪

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/thread/JobRegistryMonitorHelper.java
@@ -1,5 +1,6 @@
 package com.xxl.job.admin.core.thread;
 
+import com.xxl.job.admin.core.model.XxlJobGroup;
 import com.xxl.job.admin.core.model.XxlJobRegistry;
 import com.xxl.job.admin.core.schedule.XxlJobDynamicScheduler;
 import com.xxl.job.admin.dao.IXxlJobGroupDao;
@@ -8,13 +9,17 @@ import com.xxl.job.admin.dao.IXxlJobLogDao;
 import com.xxl.job.admin.dao.IXxlJobRegistryDao;
 import com.xxl.job.core.enums.RegistryConfig;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 
@@ -33,8 +38,6 @@ public class JobRegistryMonitorHelper {
     private final IXxlJobRegistryDao xxlJobRegistryDao;
     private final IXxlJobGroupDao xxlJobGroupDao;
 
-    private ConcurrentHashMap<String, List<String>> registMap = new ConcurrentHashMap<>();
-
     private boolean toStop = false;
 
     public void start() {
@@ -43,24 +46,42 @@ public class JobRegistryMonitorHelper {
             public void run() {
                 while (!toStop) {
                     try {
-                        // remove dead admin/executor
-                        xxlJobRegistryDao.removeDead(RegistryConfig.DEAD_TIMEOUT);
+                        // auto registry group
+                        List<XxlJobGroup> groupList = xxlJobGroupDao.findByAddressType(0);
+                        if (CollectionUtils.isNotEmpty(groupList)) {
+                            // remove dead address (admin/executor)
+                            xxlJobRegistryDao.removeDead(RegistryConfig.DEAD_TIMEOUT);
 
-                        // fresh registry map
-                        ConcurrentHashMap<String, List<String>> temp = new ConcurrentHashMap<>();
-                        List<XxlJobRegistry> list = xxlJobRegistryDao.findAll(RegistryConfig.DEAD_TIMEOUT);
-                        if (list != null) {
-                            for (XxlJobRegistry item : list) {
-                                String groupKey = makeGroupKey(item.getRegistryGroup(), item.getRegistryKey());
-                                List<String> registryList = temp.get(groupKey);
-                                if (registryList == null) {
-                                    registryList = new ArrayList<>();
+                            // fresh online address (admin/executor)
+                            Map<String, List<String>> appAddressMap = new HashMap<>();
+                            List<XxlJobRegistry> list = xxlJobRegistryDao.findAll(RegistryConfig.DEAD_TIMEOUT);
+                            if (list != null) {
+                                for (XxlJobRegistry item: list) {
+                                    if (RegistryConfig.RegistType.EXECUTOR.name().equals(item.getRegistryGroup())) {
+                                        String appName = item.getRegistryKey();
+                                        List<String> registryList = appAddressMap.get(appName);
+                                        if (registryList == null) {
+                                            registryList = new ArrayList<>();
+                                        }
+
+                                        if (!registryList.contains(item.getRegistryValue())) {
+                                            registryList.add(item.getRegistryValue());
+                                        }
+                                        appAddressMap.put(appName, registryList);
+                                    }
                                 }
-                                registryList.add(item.getRegistryValue());
-                                temp.put(groupKey, registryList);
                             }
+
+                            // fresh group address
+                            for (XxlJobGroup group: groupList) {
+                                List<String> registryList = appAddressMap.get(group.getAppName());
+                                String addressListStr = StringUtils.join(registryList, ",");
+
+                                group.setAddressList(addressListStr);
+                                xxlJobGroupDao.update(group);
+                            }
+
                         }
-                        registMap = temp;
                     } catch (Exception e) {
                         logger.error("job registry instance error:{}", e);
                     }
@@ -80,14 +101,4 @@ public class JobRegistryMonitorHelper {
         toStop = true;
         //registryThread.interrupt();
     }
-
-    private static String makeGroupKey(String registryGroup, String registryKey) {
-        return registryGroup.concat("_").concat(registryKey);
-    }
-
-    public List<String> discover(String registryGroup, String registryKey) {
-        String groupKey = makeGroupKey(registryGroup, registryKey);
-        return registMap.get(groupKey);
-    }
-
 }
